🚚 rename: supabase api export 방식 수정  (#10)
