--- src/api/supabase/index.ts
@@ -1,5 +1,0 @@
-export * from "./client";
-export * from "./authApis";
-export * from "./playlistTableAccessApis";
-export * from "./profilesTableAccessApis";
-export * from "./imageStorageAccessApis";
